Fix connection pooling (#81)
Currently we create a ClientSession for each request, which prevents
connection reuse. This moves aiohttp ClientSession to app startup and
passes the session down to the Kevel client so that we can reuse the
HTTPS connections. In local testing this reduces latency by ~200ms.

It also adds a dummy cookie jar so that we're not reusing cookies across
requests.