Bump Python version and dependencies (#82)
This is a nearly minimal diff that bumps Python to 3.11 and updates all
direct dependencies. While updating the dependencies is not strictly
necessary to move to Python 3.11, it gets rid of some deprecation
warnings and we may as well do it all in one go. We do not go to 3.12 as
not all dependencies (e.g. aiohttp) fully support it yet.

Some notable changes are:
* Moved from the deprecated circleci/python image to the new cimg/python
  image
* Moved to a slim base image, but kept the FastAPI base image for the
  time being
* Dropped the unused sentry-sdk dependency
* Dropped async/await from tests, as these were not indeed async tests,
  and failed under Python 3.11
* Moved to lifespan context manager for FastAPI, as on_event is now
  deprecated
* Renamed files in app/conf as overriding the subpackages in __init__.py
  does not work in Python 3.11